feat: add gradient visualisation to show changes in power factor
