+++ src/App.tsx
@@ -1,11 +1,11 @@
-import MandelbrotViewer from './components/MandelbrotViewer'
+import MandelbrotViewer from "./components/MandelbrotViewer";
 
 function App() {
   return (
     <div className="min-h-screen bg-gray-100">
       <MandelbrotViewer />
     </div>
-  )
+  );
 }
 
-export default App
+export default App;
